add shader model 3 support and rewrite dx9 initialization functions to add d3d9on12 support

+++ xlive/H2MOD/GUI/imgui_integration/Console/CommandsUtil.h
@@ -220,7 +220,7 @@ public:
 		}
 	}
 
-	const char* GetStringAtIndex(int headerIdx) const
+	const char* GetStringAtIndex(size_t headerIdx) const
 	{
 		assert(headerIdx < GetHeaderCount());
 		const StringLineHeader& string_header = GetHeader(headerIdx);
